requirements.txt
djangorestframework==3.15.0
Django==5.0.1
python-jose==3.3.0
python-decouple==3.8
drf-yasg==1.21.7
pillow==10.3.0
pika==1.3.2
pika-stubs==0.1.3
redis==5.0.4


